requirements.txt
aiohttp==3.9.1
asyncpg==0.29.0
beautifulsoup4==4.12.2
lxml==4.9.3
python-dotenv==1.0.0
schedule==1.2.1
loguru==0.7.2
aiofiles==23.2.1
selenium==4.15.2
webdriver-manager==4.0.1
pytz==2023.3 